Tests - TableTestBase session handling - store as attribute
Store the test session for a class on the instance to prevent starting
separate instances for helper methods. This speeds up tests and allows
lazily loading relationship values in a helper method. This also moves the
`get_value` and `get_values` to the TableTestBase class to centralize
methods.
