+++ tests/helpers.py
@@ -1,7 +1,6 @@
 import numpy as np
 from shapely.wkb import loads as load_wkb
 from shapely.wkt import loads as load_wkt
-from snowexsql.db import db_session_with_credentials
 from snowexsql.tables import MeasurementType, Observer
 
 from snowex_db.upload.layers import UploadProfileBatch, UploadProfileData
@@ -35,18 +34,6 @@ class WithUploadedFile:
         """
         attribute = getattr(table, attribute)
         return session.query(table).filter(attribute == value).all()
-
-    def get_value(self, table, attribute):
-        with db_session_with_credentials() as (engine, session):
-            obj = getattr(table, attribute)
-            result = session.query(obj).all()
-        return result[0][0]
-
-    def get_values(self, table, attribute):
-        with db_session_with_credentials() as (engine, session):
-            obj = getattr(table, attribute)
-            result = session.query(obj).all()
-        return [r[0] for r in result]
 
     def _check_metadata(self, table, attribute, expected_value):
         # Get multiple values for observers
